More efficient algorithm
The Pool Adjacent Violators Algorithm (PAVA) is still Θ(n), but
left-to-right order eliminates the `remove` calls which, in the worst
case of the right-to-left algorithm, results in O(n^2) behavior.  The
pooling in the right-to-left would still be Θ(n); the copying inherent
to moving elements to the left is the cause of O(n^2) behavior in the
worst case.  Ultimately, if it were of value, a probabilistic analysis
could be used to determine the complexity of the right-to-left
algorithm in the average case, but, seeing as there is nothing to be
gained from it, I am unlikely to carry it out. A simulation-based
approach, using non-monotonic randomly ordered vectors, reveals that
the in the average case, right-to-left is approximately twice as slow
as left-to-right (i.e ratio of average times, where average is across
30 random vectors of length 2^n, n = 2, 4, 6, ..., 22,
24). Consequently, we can safely commit right-to-left to the rubbish
heap.